requirements.txt
accelerate==0.18.0
fastapi==0.95.0
gradio==3.23
markdown2[all]==2.4.8
numpy==1.24.2
pyarrow==11.0.0
python-dotenv==1.0.0
redis==4.5.4
requests==2.28.2
sentencepiece==0.1.98
tokenizers==0.12.1
torch==2.0.0
uvicorn==0.21.1
git+https://github.com/huggingface/transformers.git@c612628045822f909020f7eb6784c79700813eda#egg=transformers